@@ -5,52 +5,68 @@
   "license": "ISC",
   "private": true,
   "scripts": {
-    "lint": "standard",
-    "build": "npm run tools && next build && next export -o dist && tsc",
+    "lint": "eslint --ext .js,.ts,.jsx .",
+    "build:api": "tsc -p server.tsconfig.json",
+    "build:web": "next build && next export -o dist",
+    "build": "npm run tools && npm run build:web && npm run build:api",
     "dev": "npm run build && npm run dev:api",
     "dev:api": "nodemon --config nodemon.json index.ts",
     "dev:web": "next",
-    "tools": "ts-node tools/gen-examples.ts",
+    "tools": "ts-node -TP server.tsconfig.json tools/gen-examples.ts",
     "now-build": "npm run build",
-    "pretest": "npm run lint",
-    "test": "tap test/*.js --reporter spec -j12",
     "start": "node dist/index.js"
   },
   "standard": {
-    "parser": "babel-eslint"
+    "parser": "@typescript-eslint/parser",
+    "plugins": [
+      "@typescript-eslint/eslint-plugin"
+    ]
   },
   "dependencies": {
-    "@sentry/integrations": "^5.4.2",
-    "@sentry/node": "^5.4.3",
-    "badgen": "^2.7.1",
-    "badgen-icons": "^0.10.1",
+    "@sentry/integrations": "^5.7.1",
+    "@sentry/node": "^5.7.1",
+    "badgen": "^3.0.1",
+    "badgen-icons": "^0.12.0",
     "byte-size": "^5.0.1",
     "cheerio": "^1.0.0-rc.3",
     "chrome-webstore": "^1.2.1",
-    "date-fns": "^1.30.1",
+    "date-fns": "^2.6.0",
     "got": "^9.6.0",
     "lodash.debounce": "^4.0.8",
     "micro": "^9.3.4",
     "millify": "^3.1.0",
-    "my-way": "^1.0.1",
+    "my-way": "^1.0.2",
     "react-debounce-render": "^5.0.0",
-    "semver": "^6.1.1",
-    "serve-handler": "^6.0.2",
-    "serve-marked": "^2.0.0"
+    "semver": "^6.3.0",
+    "serve-handler": "^6.1.2",
+    "serve-marked": "^2.0.2",
+    "url-parse": "^1.4.7"
   },
   "devDependencies": {
-    "@types/got": "^9.6.0",
+    "@types/cheerio": "^0.22.13",
+    "@types/fs-extra": "^8.0.1",
+    "@types/got": "^9.6.7",
+    "@types/lodash.debounce": "^4.0.6",
+    "@types/micro": "^7.3.3",
+    "@types/node": "^12.11.7",
+    "@types/react": "^16.9.11",
+    "@types/react-dom": "^16.9.3",
+    "@types/semver": "^6.0.2",
+    "@types/supertest": "^2.0.8",
+    "@types/url-parse": "^1.4.3",
+    "@typescript-eslint/eslint-plugin": "^2.5.0",
+    "@typescript-eslint/parser": "^2.5.0",
     "@zeit/next-typescript": "^1.1.1",
-    "babel-eslint": "^10.0.2",
-    "fs-extra": "^8.0.1",
-    "next": "^7.0.3",
-    "nodemon": "^1.19.1",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "standard": "^12.0.1",
+    "csstype": "^2.6.7",
+    "eslint": "^6.5.1",
+    "eslint-plugin-react": "^7.16.0",
+    "fs-extra": "^8.1.0",
+    "next": "^9.0.7",
+    "nodemon": "^1.19.4",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
     "supertest": "^4.0.2",
-    "tap": "^14.2.5",
-    "ts-node": "^8.3.0",
-    "typescript": "^3.5.2"
+    "ts-node": "^8.4.1",
+    "typescript": "^3.7.1-rc"
   }
 }
