@@ -1,38 +1,71 @@
-// import path from 'path'
-// import { liveBadgeList } from './badge-list'
-
-const { live: liveBadges } = require('../static/.gen/badges.json')
-
-/** not supported on Now */
-// export const liveBadges = liveBadgeList.reduce((accu, curr) => {
-//   const { meta, handlers } = require(path.resolve(__dirname, `../endpoints/${curr}`))
-//   const { title, examples, help } = meta
-//   accu[curr] = {
-//     id: curr,
-//     title,
-//     examples,
-//     routes: Object.keys(handlers),
-//     help
-//   }
-//   return accu
-// }, {})
+import path from 'path'
+import matchRoute from 'my-way'
 
 export default function genHelp (id) {
-  const meta = liveBadges.find(b => b.id === id)
+  let badgeModule
 
-  if (!meta) {
+  try {
+    badgeModule = require(path.resolve(__dirname, '../endpoints', id))
+  } catch (e) {
     return ''
   }
 
-  const { examples, routes, help } = meta
+  if (id === 'badge') {
+    return genStaticBadgeHelp(badgeModule.examples)
+  }
 
-  const Docs = `# /${id}\n\n${help || ''}`
-  const schemeLinks = routes.map(r => `- \`${r}\``)
-  const Schemes = `## Schemes\n\n${schemeLinks.join('  \n')}`
-  // @ts-ignore
-  const exampleList = Object.entries(examples)
-    .map(([url, desc]) => `- ![${url}](${url}) [${url}](${url}) <i>${desc}</i>`)
-  const Examples = `## Examples\n\n${exampleList.join('\n')}`
+  const { meta, handlers } = badgeModule
+  const { examples, help = ''} = meta
+  const routes = Object.keys(handlers)
 
-  return [Docs, Schemes, Examples].join('\n\n')
+  let md = `# /${id}\n\n${help}\n\n`
+
+  const egCats = routes.reduce((accu, curr) => {
+    accu[curr] = []
+    return accu
+  }, {})
+
+  Object.entries(examples).forEach((eg) => {
+    const scheme = routes.find(r => matchRoute(r, eg[0]))
+    if (scheme) {
+      egCats[scheme].push(eg)
+    }
+  })
+
+  md += '## Examples\n\n'
+
+  Object.entries(egCats).forEach(([cat, egs]) => {
+    // category title (scheme)
+    const hash = hashify(cat)
+    md += `\n\n<h4 id="${hash}"><a href="#${hash}">
+      <code>${cat.replace(/</g, '&lt;')}</code>
+    </a></h4>\n\n`
+
+    // category example list
+    // @ts-ignore
+    md += egs.map(egLine).join('\n')
+  })
+
+  return md
+}
+
+// turn `/github/:topic<commits|last-commit>/:owner/:repo/:ref?`
+// into `github-topic-commits-last-commit-owner-repo-ref`
+function hashify (str: string) {
+  // return str.replace(/[^\w]/g, '')
+  return str.split(/[^\w]+/).filter(Boolean).join('-')
+}
+
+function egLine ([url, desc]) {
+  return `- ![${url}](${url}) [${url}](${url}) <i>${desc}</i>`
+}
+
+function genStaticBadgeHelp (staticExamples) {
+  let md = `# Static Badge\n\n`
+
+  md += staticExamples.map(({ title, examples }) => {
+    return `### ${title}\n\n` + Object.entries(examples).map(egLine).join('\n')
+  }).join('\n\n')
+
+  return md
 }
