@@ -2,10 +2,10 @@ export default ({ data }) => {
   return data.map(({ id, title, examples }) => (
     <dl id={id} key={title}>
       <dt>
-        { id
+        {id
           ? <a className='title' href={`#${id}`}>{title}</a>
-          : <span className='title'>{title}</span> }
-        { id && <a className='doc' href={`/docs/${id}`}>?</a> }
+          : <span className='title'>{title}</span>}
+        {id && <a className='doc' href={`/${id}`}>?</a>}
       </dt>
       {
         Object.entries(examples).map(([path, desc]) => (
@@ -54,7 +54,8 @@ export default ({ data }) => {
         a.doc:hover {
           background-color: #AAA
         }
-      `}</style>
+      `}
+      </style>
     </dl>
   ))
 }
@@ -74,7 +75,7 @@ const ExampleItem = ({ desc, url }) => (
       }
       b {
         display: inline-block;
-        min-width: 15em;
+        width: 17em;
         font-family: Roboto,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif;
         text-align: right;
         font-weight: 400;
@@ -82,7 +83,7 @@ const ExampleItem = ({ desc, url }) => (
       }
       i {
         display: inline-block;
-        min-width: 210px;
+        min-width: 200px;
       }
       img {
         vertical-align: top;
@@ -103,6 +104,7 @@ const ExampleItem = ({ desc, url }) => (
           display: none;
         }
       }
-    `}</style>
+    `}
+    </style>
   </dd>
 )
