@@ -1,10 +1,16 @@
-import badgeList from '../static/.gen/badges.json'
+import badgeList from '../static/.meta/badges.json'
 
 const examples = [...badgeList.live, ...badgeList.static].reduce((accu, curr) => {
-  return accu.concat(Object.entries(curr.examples))
-}, [])
+  return (accu as any).concat(Object.entries(curr.examples))
+}, [] as [string, string][])
 
-export default ({ badgeURL, onSelect }) => {
+interface BuilderHelperProps {
+  host?: string;
+  badgeURL: string;
+  onSelect: (value: string) => void;
+}
+
+export default function BuilderHelper ({ badgeURL, onSelect }: BuilderHelperProps) {
   if (badgeURL.length < 2) {
     return <div className='helper' />
   }
@@ -12,22 +18,24 @@ export default ({ badgeURL, onSelect }) => {
   const matched = examples.filter(eg => eg[0].includes(badgeURL))
 
   const hints = matched.length === 1 && matched[0][0] === '/' + badgeURL ? null : (
-    <table><tbody>
-      {
-        matched.map(eg => (
-          <Hint
-            key={eg[0]}
-            info={eg}
-            onSelect={e => onSelect(eg[0].replace(/^\//, ''))}
-          />
-        ))
-      }
-    </tbody></table>
+    <table>
+      <tbody>
+        {
+          matched.map(eg => (
+            <Hint
+              key={eg[0]}
+              info={eg}
+              onSelect={e => onSelect(eg[0].replace(/^\//, ''))}
+            />
+          ))
+        }
+      </tbody>
+    </table>
   )
 
   return (
     <div className='helper'>
-      { hints }
+      {hints}
       <style jsx>{`
         .helper {
           height: 50vh;
@@ -41,7 +49,8 @@ export default ({ badgeURL, onSelect }) => {
           margin: 0 auto;
           padding: 1.2em 0;
         }
-      `}</style>
+      `}
+      </style>
     </div>
   )
 }
@@ -70,6 +79,7 @@ const Hint = ({ info, onSelect }) => (
         cursor: pointer;
         text-decoration: underline;
       }
-    `}</style>
+    `}
+    </style>
   </tr>
 )
