@@ -12,12 +12,12 @@ export const examples = [
     '/badge/code%20style/standard/f2a': 'code style: standard'
     }
   }, {
-    title: 'Options',
+    title: 'With Options',
     examples: {
       '/badge/icon/github?icon=github': 'use builtin icon',
       '/badge/github/github?icon': 'use builtin icon (by subject)',
-      '/badge/jquery/powered?icon=https://simpleicons.now.sh/jquery/fff': 'use external icon',
-      '/badge/icon/github?label=custom label': 'custom label',
+      '/badge/jQuery/powered?icon=https://simpleicons.now.sh/jquery/fff': 'use external icon',
+      '/badge/icon/github?label=custom%20label': 'custom label',
       '/badge/icon/github?label': 'disable label',
       '/badge/github/github?icon&label': 'use icon, disable label',
       '/badge/platform/ios,macos,tvos?list=|': 'list (custom seprator)'
