@@ -1,23 +1,25 @@
 import { useState, useEffect } from 'react'
-import BadgeExamples from '../components/badge-examples.js'
-import BadgenTitle from '../components/badgen-title.jsx'
-import Intro from '../components/home-intro.js'
-import Footer from '../components/footer.js'
-import examples from '../static/.gen/badges.json'
+import BadgeExamples from '../components/badge-examples'
+import BadgenTitle from '../components/badgen-title'
+import TopBar from '../components/top-bar'
+import Intro from '../components/home-intro'
+import Footer from '../components/footer'
+import examples from '../static/.meta/badges.json'
 
 const Index = () => {
-  const [ tab, setTab ] = useState('live')
-  const [ host, setHost ] = useState()
+  const [tab, setTab] = useState('live')
+  const [host, setHost] = useState()
   const badges = examples[tab]
 
   useEffect(() => {
-    const forceHost = new URL(window.location).searchParams.get('host')
+    const forceHost = new URL(window.location.href).searchParams.get('host')
     setHost((forceHost || window.location.origin) + '/')
   })
 
   return <>
+    <TopBar />
     <BadgenTitle host={host} />
-    <div className='docs' style={{ width: '960px' }}>
+    <div className='docs' style={{ width: '980px', margin: '0 auto' }}>
       <Intro />
       <h2 style={{ textAlign: 'center' }}>Badge Gallery</h2>
 
@@ -62,8 +64,9 @@ const Index = () => {
         color: #EEE;
         background-color: #333;
       }
-    `}</style>
-  </>
+    `}
+    </style>
+  </> // eslint-disable-line
 }
 
 export default Index
