@@ -6,7 +6,7 @@ const rel = (...args) => path.resolve(__dirname, ...args)
 
 async function main () {
   const badgeMeta = await loadBadgeMeta()
-  await fse.outputJson(rel('../static/.gen/badges.json'), badgeMeta)
+  await fse.outputJson(rel('../static/.meta/badges.json'), badgeMeta)
 }
 
 main()
